+++ src/test/java/org/sbrubbles/conditio/SignalsTest.java
@@ -8,7 +8,6 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.sbrubbles.conditio.fixtures.BasicCondition;
 import org.sbrubbles.conditio.fixtures.PleaseSignalSomethingElse;
 import org.sbrubbles.conditio.fixtures.SonOfBasicCondition;
-import org.sbrubbles.conditio.policies.HandlerNotFoundPolicy;
 import org.sbrubbles.conditio.policies.Policies;
 import org.sbrubbles.conditio.policies.ReturnTypePolicy;
 
@@ -22,11 +21,11 @@ import static org.junit.jupiter.params.provider.Arguments.arguments;
 public class SignalsTest {
   @ParameterizedTest
   @MethodSource("conditionTypeProvider")
-  @SuppressWarnings("unchecked")
+  @SuppressWarnings({"unchecked", "rawtypes"})
   public void conditionType(boolean expected, Class<? extends Condition> conditionType, Signal<?> input) {
     Predicate predicate = Signals.conditionType(conditionType);
 
-    assertEquals(expected, predicate.test((Signal) input));
+    assertEquals(expected, predicate.test(input));
   }
 
   @Test
@@ -38,7 +37,7 @@ public class SignalsTest {
   public void predicateConcatenation() {
     // the point is just that Signals' predicates can be combined and resolve to the correct type.
     // There's nothing to "run" here, per se; if this compiles, it's good
-    Predicate<Signal<BasicCondition>> p = Signals.conditionType(BasicCondition.class).and(Signals.returnType(String.class));
+    Predicate<Signal<BasicCondition>> ignored = Signals.conditionType(BasicCondition.class).and(Signals.returnType(String.class));
   }
 
   @ParameterizedTest
@@ -57,19 +56,19 @@ public class SignalsTest {
 
   static Stream<Arguments> conditionTypeProvider() {
     return Stream.of(
-      arguments(true, BasicCondition.class, withCondition(new BasicCondition(""))),
-      arguments(true, BasicCondition.class, withCondition(new SonOfBasicCondition(""))),
-      arguments(false, BasicCondition.class, withCondition(new PleaseSignalSomethingElse())),
-      arguments(false, BasicCondition.class, null)
+        arguments(true, BasicCondition.class, withCondition(new BasicCondition(""))),
+        arguments(true, BasicCondition.class, withCondition(new SonOfBasicCondition(""))),
+        arguments(false, BasicCondition.class, withCondition(new PleaseSignalSomethingElse())),
+        arguments(false, BasicCondition.class, null)
     );
   }
 
   static Stream<Arguments> returnTypeProvider() {
     return Stream.of(
-      arguments(true, String.class, withReturnType(String.class)),
-      arguments(true, Object.class, withReturnType(String.class)),
-      arguments(false, String.class, withReturnType(Object.class)),
-      arguments(false, String.class, null)
+        arguments(true, String.class, withReturnType(String.class)),
+        arguments(true, Object.class, withReturnType(String.class)),
+        arguments(false, String.class, withReturnType(Object.class)),
+        arguments(false, String.class, null)
     );
   }
 
@@ -85,8 +84,7 @@ public class SignalsTest {
     // XXX not best practice, but for these tests, should be good enough
     try (Scope scope = Scopes.create()) {
       Policies<?> policies = new Policies<>(
-        HandlerNotFoundPolicy.error(),
-        ReturnTypePolicy.expects(returnType)
+          ReturnTypePolicy.expects(returnType)
       );
 
       return new Signal<>(condition, policies, scope);