+++ src/test/java/org/sbrubbles/conditio/fixtures/logging/LoggingFixture.java
@@ -2,7 +2,6 @@ package org.sbrubbles.conditio.fixtures.logging;
 
 import org.sbrubbles.conditio.*;
 import org.sbrubbles.conditio.fixtures.AbstractFixture;
-import org.sbrubbles.conditio.policies.HandlerNotFoundPolicy;
 import org.sbrubbles.conditio.policies.Policies;
 import org.sbrubbles.conditio.policies.ReturnTypePolicy;
 import org.sbrubbles.conditio.restarts.UseValue;
@@ -42,7 +41,7 @@ public class LoggingFixture extends AbstractFixture {
 
         return scope.signal(
           getConditionProvider().apply(text),
-          new Policies<>(HandlerNotFoundPolicy.error(), ReturnTypePolicy.expects(Entry.class)),
+          new Policies<>(ReturnTypePolicy.expects(Entry.class)),
           USE_VALUE,
           RETRY_WITH);
       }